fix problem with division when computing vertical aperture

+++ source/extensions/omni.isaac.robot_engine_bridge/plugins/Sensor/CameraComponent.cpp
@@ -593,7 +593,7 @@ void CameraComponent::publishIntrinsics(std::string outputComponent,
     auto focal = pinhole.initFocal();
     // We have to ignore the vertical aperture number because our pixels are square
     // Compute it directly from the image size and horizontal aperture
-    verticalAperture = info.tex.height / info.tex.width * horizontalAperture;
+    verticalAperture = static_cast<float>(info.tex.height) / static_cast<float>(info.tex.width) * horizontalAperture;
     focal.setX(info.tex.height * focalLength / verticalAperture);
     focal.setY(info.tex.width * focalLength / horizontalAperture);
     auto center = pinhole.initCenter();
